bracketing cleanup
